fix: pyright严格模式零报错 + .gitignore完善
- pyright standard 模式 216 文件 0 错误 0 警告
- pyrightconfig.json 从 9 条压制规则减至 5 条（均为第三方类型存根问题）
- 修复 ~30 处实际代码缺陷：
  - 2 处变量可能未绑定 (agent.py timeout/summary)
  - 1 处方法参数顺序错误 (react_agent.py arun)
  - 4 处 arun_stream 返回类型不兼容
  - 7 处赋值类型不匹配 (int→float, str→Optional[str])
  - 16 处返回类型不匹配 (LLMResponse→str, SearchState→dict 等)
  - 12 处 Optional 成员访问缺少 None 检查
  - 6 处动态 Literal 类型标注问题
  - 1 处 any→Any 拼写错误 (circuit_breaker.py)
  - 1 处失效的健康检查引用 (trip.py: agent.agent→具体 agent 属性)
  - MCPTool API 使用修正 (auto_expand/_available_tools→tools)
- .gitignore: 添加 archive/output/env 变体/OS 文件等忽略规则

+++ original-hello-agents/09-上下文工程/codebase/utils.py
@@ -57,7 +57,7 @@ def ensure_dir(directory: str) -> None:
         os.makedirs(directory)
 
 
-def format_size(size_bytes: int) -> str:
+def format_size(size_bytes: float) -> str:
     """
     格式化文件大小
